// src/lib/types/activity.type.ts
export type Activity = {
  id: string;
  userId: string;
  resourceState: number;
  athleteId?: number;
  athleteResourceState?: number;
  name: string;
  distance: number;
  movingTime: number;
  elapsedTime: number;
  totalElevationGain: number;
  type: string;
  sportType: string;
  workoutType?: number | null;
  startDate: string;
  startDateLocal: string;
  timezone?: string;
  utcOffset?: number;
  locationCity?: string;
  locationState?: string;
  locationCountry?: string;
  achievementCount: number;
  kudosCount: number;
  commentCount: number;
  athleteCount: number;
  photoCount: number;
  mapId?: string;
  summaryPolyline?: string;
  mapResourceState?: number;
  trainer: boolean;
  commute: boolean;
  manual: boolean;
  private: boolean;
  visibility?: string;
  flagged: boolean;
  gearId?: string;
  startLat?: number;
  startLng?: number;
  endLat?: number;
  endLng?: number;
  averageSpeed?: number;
  maxSpeed?: number;
  hasHeartrate: boolean;
  averageHeartrate?: number | null;
  maxHeartrate?: number | null;
  heartrateOptOut: boolean;
  displayHideHeartrateOption: boolean;
  elevHigh?: number;
  elevLow?: number;
  uploadId?: string | number | null;
  uploadIdStr?: string;
  externalId?: string;
  fromAcceptedTag: boolean;
  prCount: number;
  totalPhotoCount: number;
  hasKudoed: boolean;
  createdAt?: string;
  updatedAt?: string;
  map?: {
    id?: string;
    summary_polyline?: string;
  };
};

export type StravaActivity = {
  id: number | string;
  name: string;
  distance: number;
  moving_time: number;
  elapsed_time: number;
  type: string;
  sport_type: string;
  start_date: string;
  start_date_local: string;
  total_elevation_gain?: number;
  resource_state?: number;
  achievement_count?: number;
  kudos_count?: number;
  comment_count?: number;
  athlete_count?: number;
  photo_count?: number;
  trainer?: boolean;
  commute?: boolean;
  manual?: boolean;
  private?: boolean;
  flagged?: boolean;
  has_heartrate?: boolean;
  from_accepted_tag?: boolean;
  pr_count?: number;
  total_photo_count?: number;
  has_kudoed?: boolean;
  average_speed?: number;
  max_speed?: number;
  map?: { id?: string; summary_polyline?: string; resource_state?: number };
  external_id?: string;
  upload_id?: string | number | null;
  start_latlng?: [number, number];
  end_latlng?: [number, number];
  location_city?: string;
  location_state?: string;
  location_country?: string;
  workout_type?: number | null;
  average_heartrate?: number | null;
  max_heartrate?: number | null;
  heartrate_opt_out?: boolean;
  display_hide_heartrate_option?: boolean;
  elev_high?: number;
  elev_low?: number;
  upload_id_str?: string;
  gear_id?: string;
  visibility?: string;
  athlete?: { id?: number; resource_state?: number };
  created_at?: string;
  updated_at?: string;
};